+++ kohii-exoplayer/src/main/java/kohii/v1/exoplayer/StyledPlayerViewProvider.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 Nam Nguyen, nam@ene.im
+ * Copyright (c) 2023 Nam Nguyen, nam@ene.im
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,13 +18,12 @@ package kohii.v1.exoplayer
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import com.google.android.exoplayer2.ui.PlayerView
+import com.google.android.exoplayer2.ui.StyledPlayerView
 import kohii.v1.core.Playback
 import kohii.v1.core.ViewRendererProvider
 import kohii.v1.media.Media
 
-@Deprecated(message = "PlayerView is deprecated. Use the StyledPlayerViewBridge instead.")
-class PlayerViewProvider : ViewRendererProvider() {
+class StyledPlayerViewProvider : ViewRendererProvider() {
 
   override fun getRendererType(
     container: ViewGroup,
@@ -33,17 +32,17 @@ class PlayerViewProvider : ViewRendererProvider() {
     // Note: we want to use SurfaceView on API 24 and above. But reusing SurfaceView doesn't seem to
     // be straight forward, as it is not trivial to clean the cache of old video ...
     return if (media.mediaDrm != null /* || Build.VERSION.SDK_INT >= 24 */) {
-      R.layout.kohii_player_surface_view
+      R.layout.kohii_styled_player_surface_view
     } else {
-      R.layout.kohii_player_texture_view
+      R.layout.kohii_styled_player_texture_view
     }
   }
 
   override fun createRenderer(
     playback: Playback,
     rendererType: Int
-  ): PlayerView {
+  ): StyledPlayerView {
     return LayoutInflater.from(playback.container.context)
-      .inflate(rendererType, playback.container, false) as PlayerView
+      .inflate(rendererType, playback.container, false) as StyledPlayerView
   }
 }