+++ kohii-sample/src/main/java/kohii/v1/sample/ui/youtube/YouTubeViewModel.kt
@@ -14,21 +14,53 @@
  * limitations under the License.
  */
 
-package kohii.v1.sample.youtube.data
+package kohii.v1.sample.ui.youtube
 
-import androidx.lifecycle.MutableLiveData
-import androidx.paging.DataSource
+import android.app.Application
+import androidx.lifecycle.AndroidViewModel
+import androidx.lifecycle.viewModelScope
+import androidx.paging.PagingData
+import androidx.paging.cachedIn
+import com.google.api.client.http.javanet.NetHttpTransport
+import com.google.api.client.json.gson.GsonFactory
 import com.google.api.services.youtube.YouTube
 import com.google.api.services.youtube.model.Video
-import java.util.concurrent.Executor
+import kohii.v1.sample.BuildConfig
+import kohii.v1.sample.DemoApp
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.filterNotNull
+import kotlinx.coroutines.flow.flatMapLatest
 
-class YouTubeDataSourceFactory(
-  private val apiKey: String,
-  val youtube: YouTube,
-  private val playlistId: String,
-  private val executor: Executor,
-  private val pageSize: Long
-) : DataSource.Factory<String, Video>() {
+class YouTubeViewModel(application: Application) : AndroidViewModel(application) {
+
+  private val jsonFactory = GsonFactory.getDefaultInstance()
+  private val httpTransport = NetHttpTransport()
+  private val youtube: YouTube = YouTube.Builder(
+    /* transport = */ httpTransport,
+    /* jsonFactory = */ jsonFactory,
+    /* httpRequestInitializer = */ null
+  )
+    .setApplicationName("Kohii + Youtube, " + BuildConfig.VERSION_NAME)
+    .build()
+
+  private val repository = YouTubePlaylistRepository(
+    apiKey = (application as DemoApp).youtubeApiKey,
+    youtube = youtube
+  )
+
+  private val input = MutableStateFlow<String?>(null)
+
+  @OptIn(ExperimentalCoroutinesApi::class)
+  val pagingData: Flow<PagingData<Video>> = input
+    .filterNotNull()
+    .flatMapLatest { repository.playlist(it) }
+    .cachedIn(viewModelScope)
+
+  fun load(id: String) {
+    input.value = id
+  }
 
   companion object {
     // Android & Play at Google I/O 2019
@@ -47,12 +79,5 @@ class YouTubeDataSourceFactory(
     const val YOUTUBE_VIDEOS_FIELDS =
       "items(id,snippet(title,description,thumbnails/medium,thumbnails/maxres,channelTitle))"
     // selector specifying which fields to include in a partial response.
-  }
-
-  val sourceLiveData = MutableLiveData<PageKeyedPlaylistDataSource>()
-  override fun create(): DataSource<String, Video> {
-    val source = PageKeyedPlaylistDataSource(apiKey, youtube, playlistId, executor, pageSize)
-    sourceLiveData.postValue(source)
-    return source
   }
 }